run curl in main context
